Use error-stack-parser to make better Javascript stack traces with Python error messages
