+++ src/js/package.json
@@ -16,14 +16,17 @@
   },
   "license": "Apache-2.0",
   "devDependencies": {
-    "prettier": "^2.2.1",
-    "terser": "^5.7.0",
     "rollup": "^2.48.0",
     "rollup-plugin-terser": "^7.0.2",
-    "tsd": "^0.15.1",
-    "typescript": "^4.2.4",
+    "@rollup/plugin-node-resolve": "^13.1.3",
+    "@rollup/plugin-commonjs": "^21.0.1",
+    "@types/emscripten": "^1.39.5",
+    "error-stack-parser": "^2.0.6",
     "mocha": "^9.0.2",
-    "@types/emscripten": "^1.39.5"
+    "prettier": "^2.2.1",
+    "terser": "^5.7.0",
+    "tsd": "^0.15.1",
+    "typescript": "^4.2.4"
   },
   "type": "module",
   "scripts": {
